Extract SDK helper fucntions from client
